ref(processor): Move attachement related code to sub-module
